feat: 레슨 등록시 중복된 레슨 체크

+++ badminton/src/lesson/lesson.service.ts
@@ -1,5 +1,10 @@
-import { BadRequestException, Injectable } from '@nestjs/common';
+import {
+  BadRequestException,
+  Injectable,
+  NotFoundException,
+} from '@nestjs/common';
 import { CoachRepository } from '../coach/coach.repository';
+import { LessonScheduleRequest } from './dto/lesson-schedule.request';
 import { RegisterRequest } from './dto/register.request';
 import { Lesson } from './entities/lesson.entity';
 import { LessonRepository } from './lesson.repository';
@@ -10,6 +15,15 @@ export class LessonService {
     private readonly coachRepository: CoachRepository,
     private readonly lessonRepository: LessonRepository,
   ) {}
+
+  async findLessonSchedules(dto: LessonScheduleRequest) {
+    const coach = await this.coachRepository.findOneBy({ id: dto.coachId });
+    if (!coach) {
+      throw new BadRequestException('등록되지 않은 코치입니다.');
+    }
+
+    const existLessons = await this.lessonRepository.findBy({ coachId: coach.id });
+  }
 
   async addLesson(dto: RegisterRequest) {
     const lesson = dto.toEntity();
@@ -26,10 +40,28 @@ export class LessonService {
     if (lesson.isInvalidLessonHour()) {
       throw new BadRequestException('예약할 수 없는 시간입니다.');
     }
+
+    const duplicateLesson = await this.lessonRepository.findOneBy({
+      coachId: coach.id,
+      dayOfWeek: lesson.dayOfWeek,
+      startHour: lesson.startHour,
+    });
+
+    if (duplicateLesson) {
+      throw new BadRequestException('이미 예약된 레슨 시간입니다.');
+    }
   }
 
   async removeLesson(id: number) {
-    const lesson = await this.lessonRepository.findOneByOrFail({ id });
+    const lesson = await this.lessonRepository.findOneBy({ id });
+    if (!lesson) {
+      throw new NotFoundException('등록되지 않은 레슨입니다.');
+    }
+
     await this.lessonRepository.remove(lesson);
+  }
+
+  private allSchedulesDuringWeek() {
+    
   }
 }
